refactor(grammar): collapse mixin chain, add constraint-only smoke
Grammar logits processors now derive directly from BaseLogitsProcessor
without the LogitsProcessorErrorReporter mixin. ReasoningGrammar/Grammar/
SpecLogitsProcessor share state through LogitsProcessorStates; rollback
and drain paths centralize on GenerateStream. LogitsProcessorFactory
returns a typed exception via LogitsProcessorException.h instead of the
old ad-hoc ErrorCode mapping.

Smoke: add grammar_constraint_only flag to TaskInfo. When set, the
OpenaiComparer validates each actual content against
query.response_format (json_schema or regex) and mirrors only the
non-deterministic fields (content, reasoning_content, timing, usage,
aux_info) from actual to expect. Adds a complex support_incident case
to q_r_mtp_grammar.json (nested object + 3 enums + 3 patterns + bounded
array + additionalProperties:false at every level) so an unconstrained
model output cannot pass; constraint failure leaves expect untouched
so the diff exposes the bug.

+++ rtp_llm/cpp/models/logits_processor/ThinkModeLogitsProcessor.cc
@@ -296,11 +296,11 @@ int64_t ThinkModeLogitsProcessor::acceptedTokenLen() const {
 
 int ThinkModeLogitsProcessor::tryAcceptAndFillBitmask(const SpecLogitsProcessorRequest& request) {
     if (!spec_eligible_ || request.propose_step <= 0 || request.bitmask_cpu_out == nullptr) {
-        return request.propose_step;
+        return static_cast<int>(request.propose_step);
     }
     auto snapshot = std::atomic_load_explicit(&spec_snapshot_, std::memory_order_acquire);
     if (!snapshot || !snapshot->eligible)
-        return request.propose_step;
+        return static_cast<int>(request.propose_step);
 
     StreamThinkInfo state = snapshot->info.copy();
     int             cap   = request.propose_step;
@@ -354,7 +354,7 @@ int ThinkModeLogitsProcessor::tryAcceptAndFillBitmask(const SpecLogitsProcessorR
         }
         commitToken(state, draft_token);
     }
-    return cap;
+    return static_cast<int>(cap);
 }
 
 ThinkModeLogitsProcessorPtr ThinkModeLogitsProcessor::fromGenerateInput(std::shared_ptr<GenerateInput> generate_input,
